// src/lib/api/generated/models/mediaControllerFindAllSortBy.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Beauty Pack Recommendation API
 * Backend API for beauty-pack personalization, recommendations, COD orders, and administration
 * OpenAPI spec version: 1.0.0
 */

export type MediaControllerFindAllSortBy = typeof MediaControllerFindAllSortBy[keyof typeof MediaControllerFindAllSortBy];


export const MediaControllerFindAllSortBy = {
  createdAt: 'createdAt',
  updatedAt: 'updatedAt',
  bytes: 'bytes',
  originalName: 'originalName',
} as const;