// src/lib/api/generated/models/mediaControllerFindAllSortOrder.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Beauty Pack Recommendation API
 * Backend API for beauty-pack personalization, recommendations, COD orders, and administration
 * OpenAPI spec version: 1.0.0
 */

export type MediaControllerFindAllSortOrder = typeof MediaControllerFindAllSortOrder[keyof typeof MediaControllerFindAllSortOrder];


export const MediaControllerFindAllSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;